+++ internal/parser/zz_ast_test.go
@@ -444,3 +444,17 @@ func TestSqrExpr_Pos(t *testing.T) {
 	testutils.Equal(t, "column", col, 4)
 	testutils.Equal(t, "token", tok, "SQR")
 }
+
+func TestTabExpr_Pos(t *testing.T) {
+	expr := &TabExpr{
+		Expr:   &NumberLiteral{Value: 10},
+		Line:   12,
+		Column: 4,
+		Token:  "TAB",
+	}
+
+	line, col, tok := expr.Pos()
+	testutils.Equal(t, "line", line, 12)
+	testutils.Equal(t, "column", col, 4)
+	testutils.Equal(t, "token", tok, "TAB")
+}