+++ internal/parser/zz_debug_test.go
@@ -31,6 +31,21 @@ func TestDumpExpr(t *testing.T) {
 			expected: "Ident X\n",
 		},
 		{
+			name: "TabExpr",
+			expr: &TabExpr{
+				Expr: &NumberLiteral{
+					Value:  16,
+					Line:   4,
+					Column: 2,
+					Token:  "16",
+				},
+				Line:   1,
+				Column: 1,
+				Token:  "16",
+			},
+			expected: "TAB\n  Number 16\n",
+		},
+		{
 			name: "SqrExpr",
 			expr: &SqrExpr{
 				Expr: &NumberLiteral{